Extract search and stats into src/core/search.ts
Move pending/approved/history/audit search and moderator stats into a
dedicated module. Typecheck, 167 tests, and build all pass.
